requirements.txt
black==24.10.0
click==8.1.8
colorama==0.4.6
mypy-extensions==1.0.0
packaging==24.2
pathspec==0.12.1
platformdirs==4.3.6
fastapi==0.100.0
uvicorn==0.23.0
playwright==1.39.0
openai==0.28.1
python-dotenv==1.0.0
pydantic==1.10.8
pytest==7.4.0
loguru==0.7.0
